chore: sync bindings with impyla (#67)
Replace the existing thrift definitions (with unclear lineage) with a
subset of the definitions in impyla - the official Python driver . The
definitions were copied from
https://github.com/cloudera/impyla/tree/550fa21/impala/thrift (HEAD of
master branch at this time). Those definitions are itself subset of the
server definitions at
https://github.com/apache/impala/tree/master/common/thrift .

Only relevant .thrift files and services methods are included to
minimize volume and complexity of generated code. There are no
modifications - only deletions.

Testing showed that the code remains compatible with Impala 3.