fix: fix query behavior to be predictably lazy (#77)
Ensures that Query methods
fetch rows only when Rows.Next()
is called so it is easier to cancel
statements that take too long.
Document the Exec and Query behavior.

Other docs improvements also added.

The .next.version bump is not because of this fix but a previous PR that
extended Decimal support.